% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcorge.R
\name{mrcorge}
\alias{mrcorge}
\title{Perform Mendelian Randomization based on the core gene hypothesis for polygenic exposures}
\usage{
mrcorge(
  harmonized_data,
  rank = "beta",
  K = 5,
  method_list = c("mr_ivw", "mr_weighted_mode", "mr_weighted_median"),
  seed = 1
)
}
\arguments{
\item{harmonized_data}{Data frame containing the harmonized data from the TwoSampleMR package.}

\item{rank}{Character, method to rank instruments, one of 'beta' (absolute per-allele effect size), 'h2' (per-variant heritability), 'pval' (significance), or 'zscore' (sample size-normalized z-score). Default is 'beta'.}

\item{K}{Integer, number of instrument groups. Default is 5.}

\item{method_list}{Character vector, list of MR methods to apply. Default is c("mr_ivw", "mr_weighted_mode", "mr_weighted_median").}

\item{seed}{Integer, seed for random number generation. Default is 1.}
}
\value{
A list containing:
\describe{
  \item{mrcorge}{The main MR results.}
  \item{result}{All MR results including group-specific and cumulative estimates.}
  \item{instrument}{Data frame of instruments with their respective groups.}
  \item{Q}{Q statistics for heterogeneity.}
  \item{I2}{I2 statistics for heterogeneity.}
  \item{het_P}{P-values for heterogeneity tests.}
}
}
\description{
This function performs Mendelian Randomization (MR) analyses based on the core gene hypothesis for polygenic exposures.
MR Corge identifies a small number of putative core instruments that are more likely to affect genes with a direct biological role 
in an exposure and obtains causal effect estimates based on these instruments, thereby reducing the risk of horizontal pleiotropy.
}
\examples{
# Example usage:
library(MRCorge)
data("HDL_CAD")
res <- mrcorge(HDL_CAD, rank='beta', K=20, method_list = c("mr_ivw"))
}
