% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcorge.R
\name{plot_mrcorge}
\alias{plot_mrcorge}
\title{Plot MR CORGE Results}
\usage{
plot_mrcorge(mrcorge_result, scale = "linear")
}
\arguments{
\item{mrcorge_result}{List, result from the mrcorge function.}

\item{scale}{Character, scale for the plot, either "linear" or "exp". Default is "linear".}
}
\value{
ggplot object, the plot of MR sensitivity analysis results.
}
\description{
This function plots the results of MR CORGE.
}
\examples{
# Example usage:
library(MRCorge)
data("HDL_CAD")
res <- mrcorge(HDL_CAD, rank='beta', K=20, method_list = c("mr_ivw"))
plot_mrcorge(res, scale='exp')
}
