% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcorge.R
\name{Qstatistic}
\alias{Qstatistic}
\title{Calculate Q Statistic for Heterogeneity}
\usage{
Qstatistic(b_group, se_group, b_overall)
}
\arguments{
\item{b_group}{Numeric vector, effect sizes for each group.}

\item{se_group}{Numeric vector, standard errors for each group.}

\item{b_overall}{Numeric, overall effect size.}
}
\value{
Numeric, Q statistic for heterogeneity.
}
\description{
This function calculates the Q statistic for heterogeneity in Mendelian Randomization analysis.
}
